moved components into the components folder form pages
