+++ app/client/app.js
@@ -8,7 +8,7 @@ import { hotjar } from 'react-hotjar'
 
 import { configureStore, Root } from 'pubsweet-client'
 
-import * as AuthorDetailsSchema from '@elifesciences/component-submission/client/pages/SubmissionWizard/steps/Author/schema'
+import * as AuthorDetailsSchema from '@elifesciences/component-submission/client/components/steps/Author/schema'
 import theme, { GlobalStyle } from '@elifesciences/elife-theme'
 
 import Routes from './routes'
