+++ packages/component-submission/client/components/ProgressBar.js
