+++ packages/component-submission/client/components/ThankYou.js
