+++ packages/component-submission/client/components/ThankYou.md
