+++ packages/component-submission/client/components/steps/Author/schema.js
