+++ packages/component-submission/client/components/steps/Disclosure/schema.js
