+++ packages/component-submission/client/components/steps/Editors/index.js
