+++ packages/component-submission/client/components/steps/Editors/schema.js
