+++ packages/component-submission/client/components/steps/Files/CoverLetterEditor.js
