+++ packages/component-submission/client/components/steps/Files/ManuscriptUpload.js
