+++ packages/component-submission/client/components/steps/Files/schema.js
