+++ packages/component-submission/client/components/steps/Files/utils.js
