+++ packages/component-submission/client/pages/SubmissionWizard/index.js
@@ -1,19 +1,19 @@
 import React from 'react'
 import { Switch, Redirect } from 'react-router-dom'
 import { Subscription } from 'react-apollo'
-import SubmissionOperations from './SubmissionOperations'
-import AuthorPageContainer from './steps/Author'
-import FilesPageContainer from './steps/Files'
-import SubmissionPage from './steps/Submission'
-import EditorsPageContainer from './steps/Editors'
-import DisclosurePage from './steps/Disclosure'
-import { schema as authorPageSchema } from './steps/Author/schema'
-import { schema as filesPageSchema } from './steps/Files/schema'
-import { schema as submissionPageSchema } from './steps/Submission/schema'
-import { schema as editorsPageSchema } from './steps/Editors/schema'
-import { schema as disclosurePageSchema } from './steps/Disclosure/schema'
-import WizardStep from './WizardStep'
-import { ON_UPLOAD_PROGRESS } from './operations'
+import SubmissionOperations from '../../components/SubmissionOperations'
+import AuthorPageContainer from '../../components/steps/Author'
+import FilesPageContainer from '../../components/steps/Files'
+import SubmissionPage from '../../components/steps/Submission'
+import EditorsPageContainer from '../../components/steps/Editors'
+import DisclosurePage from '../../components/steps/Disclosure'
+import { schema as authorPageSchema } from '../../components/steps/Author/schema'
+import { schema as filesPageSchema } from '../../components/steps/Files/schema'
+import { schema as submissionPageSchema } from '../../components/steps/Submission/schema'
+import { schema as editorsPageSchema } from '../../components/steps/Editors/schema'
+import { schema as disclosurePageSchema } from '../../components/steps/Disclosure/schema'
+import WizardStep from '../../components/WizardStep'
+import { ON_UPLOAD_PROGRESS } from '../../components/operations'
 import TrackedRoute from '../../../../../app/client/trackedRoute'
 import ErrorPage from '../../../../../app/client/pages/ErrorPage'
