+++ packages/component-submission/client/pages/ThankYou/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import gql from 'graphql-tag'
 import { Query } from 'react-apollo'
 import Loading from '@elifesciences/component-elife-ui/client/atoms/Loading'
-import ThankYou from './ThankYou'
+import ThankYou from '../../components/ThankYou'
 import ErrorPage from '../../../../../app/client/pages/ErrorPage'
 
 const MANUSCRIPT_TITLE = gql`
